some bugs fixed and acm in progress

+++ test/transfer/datatrans.sh
@@ -1,8 +1,9 @@
 #!/bin/bash -e
-cd /home/jiakai/programming/orzoj/server-judge/test-server/data
+cd /srv/orzoj-new-server-judge/test-server/data
 while read file t
 do
 	echo "$t..."
+
 	[ -e $t ] && rm -rf $t
 
 	7z x $file > /dev/null
@@ -10,11 +11,16 @@ do
 	cd $t
 
 	echo '<orzoj-prob-conf version="1.0">
-	<verifier standard="1"></verifier> ' > probconf.xml
+	<verifier standard="1" /> ' > probconf.xml
 
 	for cfg in `ls *.cfg | sort -n`
 	do
 		id=`echo $cfg | sed -e 's/\.cfg//g'`
+
+		if [ ! -e $id.in ]
+		then
+			continue
+		fi
 
 		echo `cat $cfg` > tmp
 		while [ 1 ]
@@ -24,17 +30,15 @@ do
 		done < tmp
 		rm tmp
 
+		cfg_time=`echo $cfg_time | sed -e 's/\\r//g'`
+		cfg_mem=`echo $cfg_mem | sed -e 's/\\r//g'`
+		cfg_score=`echo $cfg_score | sed -e 's/\\r//g'`
+
 		cfg_time=$(( $cfg_time * 1000 ))
 		cfg_mem=$(( $cfg_mem * 1024 ))
 
 		echo "
-		<case>
-			<input>$t-$id.in</input>
-			<output>$t-$id.out</output>
-			<time>$cfg_time</time>
-			<mem>$cfg_mem</mem>
-			<score>$cfg_score</score>
-		</case>
+		<case input=\"$t-$id.in\" output=\"$t-$id.out\" time=\"$cfg_time\" mem=\"$cfg_mem\" score=\"$cfg_score\" />
 		" >> probconf.xml
 
 		rm $cfg
